refactor #6. service단 로직 수정
